+++ crates/sqlite-store/src/note/filters.rs
@@ -170,10 +170,11 @@ pub(super) fn note_filter_input_notes_condition(filter: &NoteFilter) -> (String,
         },
         NoteFilter::Consumed => {
             format!(
-                "(state_discriminant in ({}, {}, {}))",
+                "(state_discriminant in ({}, {}, {}, {}))",
                 InputNoteState::STATE_CONSUMED_AUTHENTICATED_LOCAL,
                 InputNoteState::STATE_CONSUMED_UNAUTHENTICATED_LOCAL,
-                InputNoteState::STATE_CONSUMED_EXTERNAL
+                InputNoteState::STATE_CONSUMED_EXTERNAL,
+                InputNoteState::STATE_CONSUMED_EXTERNAL_ERASED
             )
         },
         NoteFilter::Expected => {